feat: adding path before acc roles prefix
